# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://stacks-node-api.mainnet.stacks.co/
    description: Mainnet
  - url: https://stacks-node-api.testnet.stacks.co/
    description: Testnet
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks 2.0 Blockchain API
  version: '1.0.0'
  description: |
    This is the documentation for the Stacks 2.0 Blockchain API.

    It is comprised of two parts; the Stacks Blockchain API and the Stacks Core API.

...
paths:
  /extended/v1/faucets/stx:
    parameters:
      - name: address
        in: query
        description: STX address
        required: true
        schema:
          type: string
      - name: stacking
        in: query
        description: Request the amount of STX needed for stacking
        required: false
        schema:
          type: boolean
          default: false
    post:
      summary: Get STX tokens
      description: Get STX tokens for the testnet
      tags:
        - Faucets
      operationId: run_faucet_stx
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/faucet/run-faucet.schema.json
              example:
                $ref: ./api/faucet/run-faucet.example.json
        500:
          description: Failed to run faucet
  /extended/v1/faucets/btc:
    parameters:
      - name: address
        in: query
        description: BTC address
        required: true
        schema:
          type: string
    post:
      summary: Get BTC tokens
      description: Get BTC tokens for the testnet
      tags:
        - Faucets
      operationId: run_faucet_btc
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/faucet/run-faucet.schema.json
              example:
                $ref: ./api/faucet/run-faucet.example.json
        500:
          description: Failed to run faucet
...